Add cargo features "client" and "server"
Since many consumers of this library do not need both client and server
features, this allows them to shrink their dependencies somewhat.
