+++ tokio-rustls/Cargo.toml
@@ -15,13 +15,16 @@ edition = "2018"
 tokio = "0.2.0"
 futures-core = "0.3.1"
 rustls = "0.17"
-webpki = "0.21"
+webpki = { version = "0.21", optional = true }
 
 bytes = { version = "0.5", optional = true }
 
 [features]
+default = ["client", "server"]
+client = ["webpki"]
 early-data = []
 dangerous_configuration = ["rustls/dangerous_configuration"]
+server = []
 unstable = ["bytes"]
 
 [dev-dependencies]
@@ -29,3 +32,11 @@ tokio = { version = "0.2.0", features = ["macros", "net", "io-util", "rt-core", 
 futures-util = "0.3.1"
 lazy_static = "1"
 webpki-roots = "0.19"
+
+[[test]]
+name = "test"
+required-features = ["client", "server"]
+
+[[test]]
+name = "badssl"
+required-features = ["client"]
